+++ src/app/core/@ngrx/products/index.ts
@@ -1,0 +1,5 @@
+export * from './products.actions';
+export * from './products.effects';
+export * from './products.selectors';
+export * from './products.state';
+export * from './products.reducer';